Create README.md
Americans in today times heavily rely on fast food or prepackaged meals due to their convenience.  Americans need to format a new way to structure their unhealthy growing obesity and diabetes problem. There are additional problems due to expensive health insurance prices and the economic unemployment of the pandemic. People need a way to optimize their food diets to push through these hard times. Do the foods in groceries and prepackaged meals have all the necessary nutrients that allow people to have and maintain a healthy lifestyle?  A possible way to fix this problem is to utilize the data set provided by the USDA that contains information on thousands of ingredients/meals and analyze the information provided to create an algorithm to sort food by category, refine and filter foods with the most frequent ingredients among the data, find foods/packages with the healthiest and unhealthiest ingredients. In fact, to expose the products that are high in energy, saturated fat, sugar and salt that can help find the optimal meal for people. We will accomplish this by using MATLAB. MATLAB has several functions to look among the most frequent ingredients in each of the packages under each specific branded food category. This helps organizing the data to become more manageable.  Since the team are relatively new to MATLAB, there are hurdles when it comes to creating the right code to execute certain tasks. Also figuring out how to implement Excel sheets with MATLAB’s format creates another challenge of its own. Using the USDA database/guidelines also help to filter out foods with either good/bad ingredients (like sugars, salts, and oils). By using these tools and combining them together, we can create a solution to a problem that can help educate and help out the growing issue.